Cleanup, database connection, further templating
removed old temporary code from files or moved them to where they are
supposed to be. Also managed to get SQLite to work with basic
write-to-db and red-from-db. Also changed the way templates are handled.
Each controller should be able to have its own template, but I still
have to improve this.